requirements.txt
Pillow==10.3.0
streamlit==1.43.2
numpy==1.22.1
opencv-python-headless==4.7.0.72